Codechange: move StringParameters to strings_internal.h
