+++ src/industry_cmd.cpp
@@ -25,7 +25,7 @@
 #include "newgrf_industrytiles.h"
 #include "autoslope.h"
 #include "water.h"
-#include "strings_func.h"
+#include "strings_internal.h"
 #include "window_func.h"
 #include "vehicle_func.h"
 #include "sound_func.h"
